chore(nodes): tidy up the node libs
