+++ internal/definition/v1/nodes/node.go
@@ -1,10 +1,7 @@
 package nodes
 
 import (
-	"crypto/sha256"
-	"encoding/hex"
 	"fmt"
-	"net/url"
 	"slices"
 	"sync"
 	"sync/atomic"
@@ -12,10 +9,7 @@ import (
 	"github.com/bigstack-oss/cube-cos-api/internal/definition/v1/base"
 	"github.com/bigstack-oss/cube-cos-api/internal/definition/v1/licenses"
 	"github.com/bigstack-oss/cube-cos-api/internal/definition/v1/metric"
-	"github.com/bigstack-oss/cube-cos-api/internal/definition/v1/settings"
 	"github.com/bigstack-oss/cube-cos-api/internal/definition/v1/status"
-	"github.com/bigstack-oss/cube-cos-api/internal/definition/v1/support"
-	"github.com/bigstack-oss/cube-cos-api/internal/definition/v1/triggers"
 	log "go-micro.dev/v5/logger"
 	"go-micro.dev/v5/registry"
 )
@@ -33,242 +27,30 @@ var (
 )
 
 type Node struct {
-	Id                string             `json:"id" yaml:"id"`
-	SerialNumber      string             `json:"serialNumber" yaml:"serialNumber"`
-	DataCenter        string             `json:"dataCenter" yaml:"dataCenter"`
-	Hostname          string             `json:"hostname" yaml:"hostname"`
-	Role              string             `json:"role" yaml:"role"`
-	Protocol          string             `json:"protocol,omitempty" yaml:"protocol,omitempty" bson:"protocol,omitempty"`
-	Address           string             `json:"address" yaml:"address"`
-	Ip                string             `json:"ip" yaml:"ip"`
-	ManagementIP      string             `json:"managementIP" yaml:"managementIP"`
-	StorageIP         string             `json:"storageIP" yaml:"storageIP"`
-	License           licenses.License   `json:"license" yaml:"license,omitempty" bson:"license,omitempty"`
-	Status            string             `json:"status" yaml:"status"`
-	CpuSpec           string             `json:"cpuSpec" yaml:"cpuSpec" bson:"cpuSpec"`
-	NetworkInterfaces []NetworkInterface `json:"networkInterfaces" yaml:"networkInterfaces" bson:"networkInterfaces"`
-	BlockDevices      []BlockDevice      `json:"blockDevices" yaml:"blockDevices" bson:"blockDevices"`
-	Vcpu              metric.Compute     `json:"vcpu" yaml:"vcpu" bson:"vcpu"`
-	Memory            metric.Space       `json:"memory" yaml:"memory" bson:"memory"`
-	Storage           metric.Space       `json:"storage" yaml:"storage" bson:"storage"`
-	UptimeSeconds     float64            `json:"uptimeSeconds" yaml:"uptimeSeconds" bson:"uptimeSeconds"`
-	Labels            map[string]string  `json:"labels,omitempty" yaml:"labels,omitempty" bson:"labels,omitempty"`
-}
+	Id           string `json:"id" yaml:"id"`
+	SerialNumber string `json:"serialNumber" yaml:"serialNumber"`
+	DataCenter   string `json:"dataCenter" yaml:"dataCenter"`
+	Hostname     string `json:"hostname" yaml:"hostname"`
+	Role         string `json:"role" yaml:"role"`
 
-type NetworkInterface struct {
-	Interface   string `json:"interface" yaml:"interface" bson:"interface"`
-	Label       string `json:"label" yaml:"label" bson:"label"`
-	BusIdSlaves string `json:"busIdSlaves" yaml:"busIdSlaves" bson:"busIdSlaves"`
-	Driver      string `json:"driver" yaml:"driver" bson:"driver"`
-	State       string `json:"state" yaml:"state" bson:"state"`
-	Speed       string `json:"speed" yaml:"speed" bson:"speed"`
-}
+	Protocol     string `json:"protocol,omitempty" yaml:"protocol,omitempty"`
+	Address      string `json:"address" yaml:"address"`
+	Ip           string `json:"ip" yaml:"ip"`
+	ManagementIP string `json:"managementIP" yaml:"managementIP"`
+	StorageIP    string `json:"storageIP" yaml:"storageIP"`
 
-type RawNetworkInterface struct {
-	Label       string `json:"label" yaml:"label" bson:"label"`
-	BusIdSlaves string `json:"busid" yaml:"busid" bson:"busid"`
-	Driver      string `json:"driver" yaml:"driver" bson:"driver"`
-	State       string `json:"state" yaml:"state" bson:"state"`
-	Speed       string `json:"speed" yaml:"speed" bson:"speed"`
-}
+	CpuSpec           string             `json:"cpuSpec" yaml:"cpuSpec"`
+	NetworkInterfaces []NetworkInterface `json:"networkInterfaces" yaml:"networkInterfaces"`
+	BlockDevices      []BlockDevice      `json:"blockDevices" yaml:"blockDevices"`
+	Vcpu              metric.Compute     `json:"vcpu" yaml:"vcpu"`
+	Memory            metric.Space       `json:"memory" yaml:"memory"`
+	Storage           metric.Space       `json:"storage" yaml:"storage"`
 
-type BlockDevice struct {
-	Serial       string             `json:"serial"`
-	Name         string             `json:"device" yaml:"device" bson:"device"`
-	Type         string             `json:"type" yaml:"type" bson:"type"`
-	SizeMiB      float64            `json:"sizeMiB" yaml:"sizeMiB" bson:"sizeMiB"`
-	Availability string             `json:"availability" yaml:"availability" bson:"availability"`
-	Status       status.BlockDevice `json:"status" yaml:"status" bson:"status"`
-}
+	License       licenses.License `json:"license" yaml:"license,omitempty"`
+	Status        string           `json:"status" yaml:"status"`
+	UptimeSeconds float64          `json:"uptimeSeconds" yaml:"uptimeSeconds"`
 
-// note:
-// rota is named by lsblk tool, it means rotational device like HDD
-type RawBlockDevice struct {
-	Type        string   `json:"type"`
-	Serial      string   `json:"serial"`
-	Name        string   `json:"name"`
-	Size        string   `json:"size"`
-	Rota        bool     `json:"rota"`
-	MountPoints []string `json:"mountpoints"`
-}
-
-func (r *RawBlockDevice) IsPartition() bool {
-	return r.Type == "part"
-}
-
-func (r *RawBlockDevice) IsBlock() bool {
-	return r.Type == "disk"
-}
-
-func (r *RawBlockDevice) NoMountPoints() bool {
-	return len(r.MountPoints) == 0
-}
-
-func (n *Node) GenUrl() string {
-	u := url.URL{Scheme: n.Protocol, Host: n.Address}
-	return u.String()
-}
-
-func (n *Node) GetMetricUrl(metric, view string) string {
-	u := url.URL{
-		Scheme: n.Protocol,
-		Host:   n.Address,
-		Path: fmt.Sprintf(
-			"/api/v1/datacenters/%s/metrics/%s/%s/hosts/%s",
-			n.DataCenter,
-			metric,
-			view,
-			n.Hostname,
-		),
-	}
-
-	return u.String()
-}
-
-func (n *Node) GetNodeUrl() string {
-	u := url.URL{
-		Scheme: n.Protocol,
-		Host:   n.Address,
-		Path: fmt.Sprintf(
-			"/api/v1/datacenters/%s/nodes/%s",
-			n.DataCenter,
-			n.Hostname,
-		),
-	}
-
-	return u.String()
-}
-
-func (n *Node) PostLicenseUrl() string {
-	u := url.URL{
-		Scheme: n.Protocol,
-		Host:   n.Address,
-		Path:   fmt.Sprintf("/api/v1/datacenters/%s/licenses/hosts/%s", base.DataCenterName, n.Hostname),
-	}
-
-	return u.String()
-}
-
-func (n *Node) GetTuningUrl() string {
-	u := url.URL{
-		Scheme:   n.Protocol,
-		Host:     n.Address,
-		Path:     fmt.Sprintf("/api/v1/datacenters/%s/tunings/parameters", n.DataCenter),
-		RawQuery: "allNodes=false",
-	}
-
-	return u.String()
-}
-
-func (n *Node) GetSettingUrl(path string) string {
-	u := url.URL{
-		Scheme:   n.Protocol,
-		Host:     n.Address,
-		Path:     path,
-		RawQuery: "clusterWise=false",
-	}
-
-	return u.String()
-}
-
-func (n *Node) GetSupportFileUrl() string {
-	u := url.URL{
-		Scheme: n.Protocol,
-		Host:   n.Address,
-		Path:   fmt.Sprintf("/api/v1/datacenters/%s/supportFiles/hosts/%s", n.DataCenter, n.Hostname),
-	}
-
-	return u.String()
-}
-
-func (n *Node) DownloadSupportFileUrl(setname, filename string) string {
-	u := url.URL{
-		Scheme: n.Protocol,
-		Host:   n.Address,
-		Path:   fmt.Sprintf("/api/v1/datacenters/%s/supportFiles/%s/%s", n.DataCenter, setname, filename),
-	}
-
-	return u.String()
-}
-
-func (n *Node) PatchTuningUrl(tuning string) string {
-	u := url.URL{}
-	u.Scheme = n.Protocol
-	u.Host = n.Address
-	u.Path = fmt.Sprintf("/api/v1/datacenters/%s/tunings/parameters/%s", base.DataCenterName, tuning)
-	return u.String()
-}
-
-func (n *Node) EnableOrDisableTuningUrl(tuning string) string {
-	u := url.URL{}
-	u.Scheme = n.Protocol
-	u.Host = n.Address
-	u.Path = fmt.Sprintf("/api/v1/datacenters/%s/tunings/parameters/%s/enable", base.DataCenterName, tuning)
-	return u.String()
-}
-
-func (n *Node) ResetTuningUrl(tuning string) string {
-	u := url.URL{}
-	u.Scheme = n.Protocol
-	u.Host = n.Address
-	u.Path = fmt.Sprintf("/api/v1/datacenters/%s/tunings/parameters/%s/reset", base.DataCenterName, tuning)
-	return u.String()
-}
-
-func (n *Node) PatchTuningTaskUrl() string {
-	u := url.URL{}
-	u.Scheme = n.Protocol
-	u.Host = n.Address
-	u.Path = fmt.Sprintf("/api/v1/datacenters/%s/tunings/tasks", base.DataCenterName)
-	return u.String()
-}
-
-func (n *Node) PatchTriggerTaskUrl(trigger triggers.ApiSchema) string {
-	u := url.URL{}
-	u.Scheme = n.Protocol
-	u.Host = n.Address
-	u.Path = fmt.Sprintf("/api/v1/datacenters/%s/triggers/tasks/%s", base.DataCenterName, trigger.Name)
-	return u.String()
-}
-
-func (n *Node) CreateSupportFileUrl(file support.File) string {
-	u := url.URL{}
-	u.Scheme = n.Protocol
-	u.Host = n.Address
-	u.Path = fmt.Sprintf("/api/v1/datacenters/%s/supportFiles", base.DataCenterName)
-	return u.String()
-}
-
-func (n *Node) PatchSupportFileTaskUrl(file support.File) string {
-	u := url.URL{}
-	u.Scheme = n.Protocol
-	u.Host = n.Address
-	u.Path = fmt.Sprintf("/api/v1/datacenters/%s/supportFiles/%s", base.DataCenterName, file.Group)
-	return u.String()
-}
-
-func (n *Node) PatchSettingTaskUrl(setting settings.Setting) string {
-	u := url.URL{}
-	u.Scheme = n.Protocol
-	u.Host = n.Address
-	u.Path = fmt.Sprintf("/api/v1/datacenters/%s/settings/tasks", base.DataCenterName)
-	return u.String()
-}
-
-func (n *Node) DeleteRepairingTaskUrl() string {
-	u := url.URL{}
-	u.Scheme = n.Protocol
-	u.Host = n.Address
-	u.Path = fmt.Sprintf("/api/v1/datacenters/%s/healths/tasks/repairing", base.DataCenterName)
-	return u.String()
-}
-
-func (n *Node) DeleteModuleRepairingTaskUrl(module string) string {
-	u := url.URL{}
-	u.Scheme = n.Protocol
-	u.Host = n.Address
-	u.Path = fmt.Sprintf("/api/v1/datacenters/%s/healths/tasks/repairing/%s", base.DataCenterName, module)
-	return u.String()
+	Labels map[string]string `json:"labels,omitempty" yaml:"labels,omitempty"`
 }
 
 func (n *Node) IsLocal() bool {
@@ -303,44 +85,20 @@ func IsLocalAddress(address string) bool {
 	return base.AdvertiseAddr == address
 }
 
-func GenerateNodeHashByMacAddr() (string, error) {
-	macAddr, err := base.GetMacAddr(base.NetMajorInterface)
-	if err != nil {
-		return "", err
+func New(node *registry.Node) Node {
+	return Node{
+		Role:         node.Metadata["role"],
+		Id:           node.Id,
+		SerialNumber: node.Metadata["serialNumber"],
+		DataCenter:   node.Metadata["dataCenter"],
+		Protocol:     node.Metadata["protocol"],
+		Ip:           node.Metadata["ip"],
+		Hostname:     node.Metadata["hostname"],
+		Address:      node.Address,
+		Labels: map[string]string{
+			"isGpuEnabled": node.Metadata["isGpuEnabled"],
+		},
 	}
-
-	hash := sha256.Sum256([]byte(macAddr))
-	return hex.EncodeToString(hash[:])[:8], nil
-}
-
-func GetNodesByRole(role string) ([]Node, error) {
-	svcs, err := GetDiscoveredServices()
-	if err != nil {
-		return nil, err
-	}
-
-	list := []Node{}
-	for _, svc := range svcs {
-		nodes := parseNodesByRole(svc, role)
-		if len(nodes) != 0 {
-			list = append(list, nodes...)
-		}
-	}
-
-	return list, nil
-}
-
-func parseNodesByRole(svc *registry.Service, role string) []Node {
-	nodes := []Node{}
-	for _, node := range svc.Nodes {
-		if node.Metadata["role"] != role {
-			continue
-		}
-
-		nodes = append(nodes, New(node))
-	}
-
-	return nodes
 }
 
 func GetDiscoveredServices() ([]*registry.Service, error) {
@@ -362,62 +120,15 @@ func GetDiscoveredServices() ([]*registry.Service, error) {
 	return svcs, nil
 }
 
-func HostnameMap() (map[string]Node, error) {
-	svcs, err := GetDiscoveredServices()
-	if err != nil {
-		return nil, err
-	}
-
-	nodeMap := map[string]Node{}
-	for _, svc := range svcs {
-		nodes := parseNodes(svc)
-		for _, node := range nodes {
-			nodeMap[node.Hostname] = node
-		}
-	}
-
-	return nodeMap, nil
-}
-
-func parseNodes(svc *registry.Service) []Node {
-	nodes := []Node{}
-
-	for _, node := range svc.Nodes {
-		if IsLocal(node.Metadata["hostname"]) {
-			continue
-		}
-
-		nodes = append(nodes, New(node))
-	}
-
-	return nodes
-}
-
-func New(node *registry.Node) Node {
-	return Node{
-		Role:         node.Metadata["role"],
-		Id:           node.Id,
-		SerialNumber: node.Metadata["serialNumber"],
-		DataCenter:   node.Metadata["dataCenter"],
-		Protocol:     node.Metadata["protocol"],
-		Ip:           node.Metadata["ip"],
-		Hostname:     node.Metadata["hostname"],
-		Address:      node.Address,
-		Labels: map[string]string{
-			"isGpuEnabled": node.Metadata["isGpuEnabled"],
-		},
-	}
-}
-
 func GetControlNodes() ([]Node, error) {
 	controllers := []Node{}
 
-	nodes, err := GetNodesByRole("control")
+	nodes, err := GetNodesByRole(RoleControl)
 	if err == nil && len(nodes) > 0 {
 		controllers = append(controllers, nodes...)
 	}
 
-	nodes, err = GetNodesByRole("control-converged")
+	nodes, err = GetNodesByRole(RoleControlConverged)
 	if err == nil && len(nodes) > 0 {
 		controllers = append(controllers, nodes...)
 	}
@@ -433,23 +144,9 @@ func GetControlNodes() ([]Node, error) {
 }
 
 func GetPeerControls() ([]Node, error) {
-	controllers := []Node{}
-
-	nodes, err := GetNodesByRole("control")
-	if err == nil && len(nodes) > 0 {
-		controllers = append(controllers, nodes...)
-	}
-
-	nodes, err = GetNodesByRole("control-converged")
-	if err == nil && len(nodes) > 0 {
-		controllers = append(controllers, nodes...)
-	}
-
-	if len(controllers) == 0 {
-		return nil, fmt.Errorf(
-			"failed to get control nodes(control or control-converged): %v",
-			err,
-		)
+	controllers, err := GetControlNodes()
+	if err != nil {
+		return nil, err
 	}
 
 	for i, controller := range controllers {
@@ -469,14 +166,6 @@ func GetController() (*Node, error) {
 	}
 
 	return &nodes[0], nil
-}
-
-func Lock() {
-	nodeList.Lock()
-}
-
-func Unlock() {
-	nodeList.Unlock()
 }
 
 func List() []Node {
@@ -563,39 +252,10 @@ func GetMap() map[string]Node {
 	return nodes
 }
 
-func GetNodesByRoles() []Node {
-	nodes := []Node{}
-	for _, role := range roles {
-		switch role {
-		case RoleControl:
-			nodes = append(nodes, Control.Load().Nodes...)
-		case RoleCompute:
-			nodes = append(nodes, Compute.Load().Nodes...)
-		case RoleStorage:
-			nodes = append(nodes, Storage.Load().Nodes...)
-		case RoleControlConverged:
-			nodes = append(nodes, ControlConverged.Load().Nodes...)
-		case RoleModerator:
-			nodes = append(nodes, Moderator.Load().Nodes...)
-		case RoleEdgeCore:
-			nodes = append(nodes, EdgeCore.Load().Nodes...)
-		}
-	}
-
-	return nodes
+func Lock() {
+	nodeList.Lock()
 }
 
-func convertToHosts(nodes []Node) []Host {
-	hosts := []Host{}
-	for _, node := range nodes {
-		hosts = append(
-			hosts,
-			Host{
-				Name: node.Hostname,
-				Ip:   node.Ip,
-			},
-		)
-	}
-
-	return hosts
+func Unlock() {
+	nodeList.Unlock()
 }
