+++ internal/operators/v1/nodes/sync.go
