+++ internal/operators/v1/settings/record.go
@@ -9,7 +9,7 @@ import (
 
 func (o *Operator) handleExit(setting settings.Setting, err error) {
 	if err != nil {
-		log.Errorf("settings: failed to %s %s: %v", setting.Status.Desired, setting.Type, err)
+		log.Errorf("settings: failed to %s %s(%v)", setting.Status.Desired, setting.Type, err)
 		setting.SetError()
 	} else {
 		log.Infof("settings: %s %s successfully", setting.Status.Desired, setting.Type)
@@ -24,7 +24,7 @@ func (o *Operator) handleExit(setting settings.Setting, err error) {
 func (o *Operator) reportToController(setting settings.Setting) {
 	node, err := nodes.GetController()
 	if err != nil {
-		log.Errorf("settings: failed to get controller nodes: %v", err)
+		log.Errorf("settings: failed to get controller nodes(%v)", err)
 		return
 	}
 
@@ -34,13 +34,13 @@ func (o *Operator) reportToController(setting settings.Setting) {
 		SetBody(setting.GenTaskUpdate()).
 		Patch(node.PatchSettingTaskUrl(setting))
 	if err != nil {
-		log.Errorf("settings: failed to send setting %s to %s: %v", setting.Type, node.Hostname, err)
+		log.Errorf("settings: failed to send setting %s to %s(%v)", setting.Type, node.Hostname, err)
 		return
 	}
 
 	if resp.IsError() {
 		log.Errorf(
-			"settings: error response from %s %s update: %v",
+			"settings: error response from %s %s update(%v)",
 			node.Hostname,
 			setting.Type,
 			string(resp.Body()),
