+++ internal/operators/v1/supportfiles/report.go
@@ -1,8 +1,6 @@
 package supportfiles
 
 import (
-	"errors"
-
 	"github.com/bigstack-oss/bigstack-dependency-go/pkg/http"
 	"github.com/bigstack-oss/cube-cos-api/internal/cubecos"
 	"github.com/bigstack-oss/cube-cos-api/internal/definition/v1/nodes"
@@ -12,7 +10,7 @@ import (
 
 func (o *Operator) handleExit(file *support.File, err error) {
 	if err != nil {
-		log.Errorf("supportfiles: failed to %s %s: %v", file.Status.Desired, file.Group, err)
+		log.Errorf("supportfiles: failed to %s %s(%v)", file.Status.Desired, file.Group, err)
 		file.SetError()
 	} else {
 		log.Infof("supportfiles: %s %s successfully", file.Status.Desired, file.Group)
@@ -20,17 +18,14 @@ func (o *Operator) handleExit(file *support.File, err error) {
 	}
 
 	cubecos.SetSupportFileComment(*file)
-	err = o.reportToController(*file)
-	if err != nil {
-		return
-	}
+	o.reportToController(*file)
 }
 
-func (o *Operator) reportToController(file support.File) error {
+func (o *Operator) reportToController(file support.File) {
 	node, err := nodes.GetController()
 	if err != nil {
-		log.Errorf("supportfiles: failed to get controller nodes: %v", err)
-		return err
+		log.Errorf("supportfiles: failed to get controller nodes(%v)", err)
+		return
 	}
 
 	h := http.GetGlobalHelper()
@@ -39,14 +34,16 @@ func (o *Operator) reportToController(file support.File) error {
 		SetBody(file.GenTaskUpdate()).
 		Patch(node.PatchSupportFileTaskUrl(file))
 	if err != nil {
-		log.Errorf("supportfiles: failed to send support file %s to %s: %v", file.Group, node.Hostname, err)
-		return err
+		log.Errorf("supportfiles: failed to send support file %s to %s(%v)", file.Group, node.Hostname, err)
+		return
 	}
 
 	if resp.IsError() {
-		log.Errorf("supportfiles: failed to send support file %s to %s: %v", file.Group, node.Hostname, string(resp.Body()))
-		return errors.New(string(resp.Body()))
+		log.Errorf(
+			"supportfiles: failed to send support file %s to %s(%v)",
+			file.Group,
+			node.Hostname,
+			string(resp.Body()),
+		)
 	}
-
-	return nil
 }
