+++ internal/operators/v1/triggers/record.go
@@ -1,8 +1,6 @@
 package triggers
 
 import (
-	"errors"
-
 	"github.com/bigstack-oss/bigstack-dependency-go/pkg/http"
 	"github.com/bigstack-oss/cube-cos-api/internal/definition/v1/nodes"
 	"github.com/bigstack-oss/cube-cos-api/internal/definition/v1/triggers"
@@ -11,7 +9,7 @@ import (
 
 func (o *Operator) handleExit(trigger triggers.ApiSchema, err error) {
 	if err != nil {
-		log.Errorf("triggers: failed to %s %s: %v", trigger.Status.Desired, trigger.Name, err)
+		log.Errorf("triggers: failed to %s %s(%v)", trigger.Status.Desired, trigger.Name, err)
 		trigger.SetError()
 	} else {
 		log.Infof("triggers: %s %s successfully", trigger.Status.Desired, trigger.Name)
@@ -23,11 +21,11 @@ func (o *Operator) handleExit(trigger triggers.ApiSchema, err error) {
 	}
 }
 
-func (o *Operator) reportToController(trigger triggers.ApiSchema) error {
+func (o *Operator) reportToController(trigger triggers.ApiSchema) {
 	node, err := nodes.GetController()
 	if err != nil {
-		log.Errorf("triggers: failed to get controller nodes: %v", err)
-		return err
+		log.Errorf("triggers: failed to get controller nodes(%v)", err)
+		return
 	}
 
 	h := http.GetGlobalHelper()
@@ -36,14 +34,16 @@ func (o *Operator) reportToController(trigger triggers.ApiSchema) error {
 		SetBody(trigger.GenTaskUpdate()).
 		Patch(node.PatchTriggerTaskUrl(trigger))
 	if err != nil {
-		log.Errorf("triggers: failed to send trigger %s to %s: %v", trigger.Name, node.Hostname, err)
-		return err
+		log.Errorf("triggers: failed to send trigger %s to %s(%v)", trigger.Name, node.Hostname, err)
+		return
 	}
 
 	if resp.IsError() {
-		log.Errorf("triggers: failed to send trigger %s to %s: %s", trigger.Name, node.Hostname, string(resp.Body()))
-		return errors.New(string(resp.Body()))
+		log.Errorf(
+			"triggers: failed to send trigger %s to %s(%s)",
+			trigger.Name,
+			node.Hostname,
+			string(resp.Body()),
+		)
 	}
-
-	return nil
 }
